fix(iroh-relay): Fix proptests, make `Datagrams::segment_size` be an `Option<NonZeroU16>` (#3404)
## Description

The proptests were failing in some rare case where
`Datagrams::segment_size` was generated as `Some(0)`. (For some reason
we haven't hit that before?).

I added the failing case hash to the proptest regressions file and fixed
the problem.

I also changed `Datagrams::segment_size` to be an `Option<NonZeroU16>`
instead of `Option<u16>`, which checks this invariant on the type level
now.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.