fix: wait for at least one ipv6 and ipv4 qad report (#3413)
## Description

If an endpoint only has a single relay URL configured in its relay map,
we currently abort the QAD net reports after a single successful report.
However, we also decide whether an endpoint supports IPv6 on the fact if
a QAD IPv6 report completed successfully or not. Those are *two* reports
though, and if we abort after the first one, we falsely assume that the
endpoint has no IPv6 connectivity.

This PR changes this in a rather simple way: If we did start both IPv4
and IPv6 reports, we always wait at least until one of each completed.

Unrelated to this change, the PR also adds logging when the best address
changes.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)

---------

Co-authored-by: Philipp Krüger <philipp.krueger1@gmail.com>