fix(iroh): Re-batch received relay datagram batches in case they exceed `max_receive_segments` (#3414)
## Description

Fixes a panic in `RelayTransport::poll_recv` with the new relay code.

Since #3389 we forward the whole GSO batch via the relay. This then gets
read on the other receiving endpoint, which tries to fill its `buf_out`,
which might not be big enough, because its size is initialized with
`max_receive_segments` taken into account, which might be different from
the sending endpoint's `max_transmit_segments`.

So in case we get a batch that's too big, we use a `DatagramReBatcher`
to split up our datagram batch into chunk sizes we can consume.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.