docs: grammar nitpick (#3383)
## Description

Really small grammar check, no breaking issues I'm aware of. "allows to"
isn't a phrase that is technically correct.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->
"To address this, the thing allows for the configuration of a discovery
service." could work too, but I prefer using "you" in instructions to
make things just the littlest bit less abstract.

## Change checklist
<!-- Remove any that are not relevant. -->
- [X] Self-review.