feat: Allow configuring the max number of TLS tickets (#3442)
## Description

Allows increasing the max number of TLS tickets without having to patch
iroh. This is needed if you want to do a large number of 0rtt
connections to different clients.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

Q: is this enough, or should we allow passing in an `Arc<dyn
ClientSessionStore>`? The latter would mean exposing one more rustls
type.

Note: this is mostly for playing around with it, not sure if we want to
merge it as is. But then, it seems pretty straightforward. If we ever
want the ability to provide an `Arc<dyn ClientSessionStore>` we can
always make the `fn max_tls_tickets` just init an in-mem provider.

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)