feat: upgrade redb to v3 compatible format (#3483)
## Description

We are using redb in iroh-dns-server. The recently released redb v3 has
incompatible changes to the database format. The recommended upgrade
procedure is to upgrade the database with `Database::upgrade` on redb
v2.6+. The upgrade has to be done on redb v2, not v3. See
https://github.com/cberner/redb/blob/master/CHANGELOG.md#removes-support-for-file-format-v2.

This updates redb to 2.6 and performs the upgrade when opening the
database. Calling upgrade on an already upgraded database is a no-op.

With this merged in the next release, we can then update redb to v3 in
the version after.


## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)