Simplify delayed discovery start
This no longer tries to figure out if discovery is needed based on how
recently we may have communicated with a remote endpoint.  Maybe
that's too simple and we'll have to add back some extra check for
discovery.  But most active connections would have had a response by
that time and not trigger the discovery.