Convert to canonical IP address in IpSender
Our AsyncUdpSocket is always an IPv6 socket, because we need to sent
to our IPv6 ULAs for the mapped addresses.  This means Quinn will
convert any IPv4 destination into an IPv4-mapped IPv6
address (::ffff:a.b.c.d).

But our sockets are bound to a specific family, so even if the OS
supports and IPv6 socket that can handle IPv4 it will not accept those
addresses.  So when we are using the destinations we need to convert
back to the canonical addresses in the IPSender.