fix(iroh): convert to canonical IP address in IpSender (#3506)
## Description

Our AsyncUdpSocket is always an IPv6 socket, because we need to sent to
our IPv6 ULAs for the mapped addresses. This means Quinn will convert
any IPv4 destination into an IPv4-mapped IPv6 address (::ffff:a.b.c.d).

But our sockets are bound to a specific family, so even if the OS
supports and IPv6 socket that can handle IPv4 it will not accept those
addresses. So when we are using the destinations we need to convert back
to the canonical addresses in the IPSender.


Based on
https://github.com/n0-computer/iroh/commit/4eb0f07e34cccbb53e1842093d29c7b013ff4f5d