feat: add `on_incoming`, that allows you to reject a connection at the incoming stage
Unforunately, this does not work as expected. Currently, you only get the mapped address on the QUIC connection, not the actual IP address used to send the message to you. This is further complicated by the fact that often our first messages are sent via the relay, meaning that even if you had the IP address of the relay, there is not much value in blocking by that IP, unless you were trying to block an entire relay.