refactor(iroh-relay)!: improve relay map api (#3563)
## Description

#3542 included unintended changes to the `RelayMap` API. This cleans
this up: We don't call entries in the relay map *endpoints*, instead we
simply call them *relays*.

Also improves the API docs and adds some examples for how to work with
the `FromIterator` generic types used in construction and access. And
renamed a couple of variables at use site in tests mostly to follow the
new pattern.

## Breaking Changes

### Renamed

* `iroh_relay::RelayMap::get_endpoint` -> `get`
* `iroh_relay::RelayMap::endpoints` -> `relays`
* `iroh_relay::RelayMap::contains_endpoint` -> `contains`

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)