Broken: add relay to a new connection that is direct
When a new connection arrives that is direct, we should add the relay
connection.  Because the initial connection was probably racing direct
and relay and direct won.

Unfortunately immediately adding this path fails because the
connection does not have have the remote CIDs to allow more paths.  So
we need some sort of retry mechanism.

This is currently broken, probably a silly bug in how the retry is
scheduled or consumed.  It's in the middle of some warn!()-style
debugging.