chore(iroh): Update wasm-bindgen version to fix CI (#3589)
## Description

Fixes the wasm CI by updating our wasm-bindgen version.

## Breaking Changes

None, these are just some chores.

## Notes & open questions

Kind of annoying that this keeps breaking. A "real fix" here would be to
somehow make sure that the wasm-bindgen-cli that we install in the CI
workflow matches the wasm-bindgen version in the Cargo.lock. That needs
some thought on how it'd work. I'd prefer to merge this fix before we do
that though, given that this blocks merging anything.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.