Don't store remote NAT candidates
Turns out we don't do anything with those.  We only use this to decide
to check if we need to holepunch again.

This calls trigger_holepunching for every connection event which will
wake it up too much, because we only really need to do it for the
connection we are using to holepunch.  But the state tracking is
easier like this and the cost probably not too high.