fix: stop relay actor on endpoint close (#3601)
## Description

Since the transport abstraction was introduced, we no longer shut down
the relay actor when calling `Endpoint::close`. The relay actor is
currently only stopped when the last clone of the `iroh::Endpoint` is
dropped. This is surprising and wasteful. This PR fixes this so that the
relay actor is stopped when calling `Endpoint::close`.

Background: Since introducing the transport abstraction, the relay actor
handle is now owned by `Transports` which is owned by `MagicUdpSocket`
which is owned by the `quinn::Endpoint` as its `AsyncUdpSocket`. The
`AsyncUdpSocket` trait has no notification when the endpoint is closing,
thus currently the relay transport (and the relay actor, which it owns)
is not notified at all when the endpoint is shutting down. Therefore, it
keeps running (and potentially reconnecting to the relay or receiving
messages) until it is stopped when the last clone of the endpoint is
dropped and thus the `AbortOnDropHandle` owning the relay actor tasks is
dropped, finally stopping the relay actor tasks.

This PR exposes an existing cancellation token within the relay actor to
the magicsocket, allowing it to cancel the relay actors when calling
`Endpoint::close`.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

I tested this by looking at logs after closing an endpoint. Not sure if
there is a good way to test this. I'd vote to merge this even without a
test because it fixes an issue, unless someone has a good idea on how to
test this.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [x] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)