Instead of duplicating the entire connection API, have a type parameter on Connection that describes the connection state, and implement the fns alpn and remote_id differently depending on what the state is.
The upside is that it should be easier to write code that works with both
a 0rtt connection and a fully initialized connection without having to
define a trait to abstract over both variants.

The downside is that it is a bit fancy with the types. But then for the
non 0rtt case you would never see the fanciness, so maybe it is OK?