fix(iroh-bench): Gracefully close the quinn benchmark (#3621)
## Description

Adds a call to `quinn::Endpoint::wait_idle` as the equivalent to
`iroh::Endpoint::close`.

This fixes the quinn benchmark not timing out, but closing gracefully
instead for me.
(Before, running `cargo run -p iroh-bench -- quinn` would take 18s to
finish due to quinn waiting for a probe timeout, after this, this bench
runs in 2s for me.)


## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.