feat: Do not force relay URLs to be absolute
It is really up to the user.  Forcing this was always a choice that
made me slightly uncomfortable.  Some weird places might need relative
URLs.  Or places that need to be able to use this same URL using
OpenSSL, which has a bug handling certificates that do not have the
absolute DNS name.

The important part is that our defaults still use an absolute DNS
name.  And we also strongly encourage it.

Fixes #3612