feat: Do not force relay URLs to be absolute (#3623)
## Description

It is really up to the user.  Forcing this was always a choice that
made me slightly uncomfortable.  Some weird places might need relative
URLs.  Or places that need to be able to use this same URL using
OpenSSL, which has a bug handling certificates that do not have the
absolute DNS name.

The important part is that our defaults still use an absolute DNS
name.  And we also strongly encourage it.

## Breaking Changes

None

## Notes & open questions

Fixes #3612, obsoletes #3487.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.