refactor: use boxed watcher, not watchable, on connection (#3632)
## Description

Alternative to https://github.com/n0-computer/iroh/pull/3631

Replaces the `Watchable`s for path changes on the `Connection` with a
boxed `Watcher`. The watcher is boxed because it would increase the
`Connection` struct size significantly otherwise because the
mapped-and-joined watcher with a `SmallVec` of `PathInfo` inside is ~600
bytes atm.

The benefit of storing a `Watcher` and not a `Watchable` is that the
watcher streams now close once the EndpointStateActor drops the state
for the connection, which it does after the connection is closed.

Also adds a test for path watching, including testing that the streams
now close when the connection closes.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)