refactor(multipath): Stop inactive endpoint actors (#3643)
## Description

Fixes #3638 (partially)

This is a first, small solution to stop inactive endpoint actors after
an idle timeout. I implemented it such that the *actor* decides once to
stop, while making sure that we *never* create senders to actors that
are shutting down.

Logic in the actor:
* The actor enters an idle timeout (set to 60 seconds) once it has no
active connections, an empty inbox, and no inbox senders
* Once the timeout expires, it is rechecked that the idle conditions
hold, and if so the actor exits
* Once any of the idle conditions don't hold anymore, the idle timeout
is deactivated and restarted once the conditions are met again

The actor checks if the inbox's sender strong count equals 1, which
means that no senders exist apart from the one held in the endpoint map.
This check is protected with a mutex, to enter a critical section for
closing the inbox while the lock is held in case the conditions are met.
This is to ensure that there cannot be a race condition where a sender
is cloned out right after the check in the actor returns true, but
before the inbox is closed.

Logic in the endpoint map:
* When handing out senders, we acquire the shared lock, and check that
the channel is not closed while the lock is held. This ensures that the
actor never closes while a sender is alive. If the actor is closed, we
remove the handle to the dead actor and create a new actor.
* On regular intervals (set to 60 seconds) the magicsock actor removes
handles to dead actors.


## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

* I *think* my logic around the critical section and ensuring that we
never close the actor while senders exist is sound. However, it needs
careful review and tests. I'll do some thinking on how to best test
this.
* Instead of employing an interval to remove dead actor handles, we
could use a channel where the actor informs an outside-task which
endpoint actors terminated, so that the outside-task can then lock the
endpoint map and remove just those. Not sure if that's worth it.
* Another solution here might be to spawn the actor tasks into a join
set in the magicsock actor. However this would need further refactoring
and would likely make spawning actors async. I think I'd prefer to keep
that sync because it makes the surrounding code a lot simpler.
* This does not yet implement some of the more advanced reasoning that
#3638 proposes. I think we should start with something simple that
prevents memory exhaustion and tweak as needed. However, it could also
be argued that we should start with a more featureful design right away.

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)