fix(iroh): Clear `EndpointStateActor::selected_path` once the last connection closes (#3650)
## Description

Addresses #3602 for multipath.

This implements the first bullet point from @flub's suggestion in the
above issue: We clear the `selected_path` once the last known connection
to an endpoint closes.

This means that a new connection attempt after that will instead send to
all addresses again, and avoids the case where we send on e.g. the old
port from a previous restart of the endpoint we're connecting to.

This turns the `test_0rtt_after_server_restart` test green.

## Notes & open questions

I *think* this will still fail if we have an "open" connection to the
server that's in the process of timing out and we open a new connection
to the restarted server while that's happening. I'm not sure though.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.